feat: create ADD_CARD modal and allow going back to previous modal
