+++ ui/component/walletTipAmountSelector/index.js
@@ -2,6 +2,7 @@ import { connect } from 'react-redux';
 import { selectBalance } from 'redux/selectors/wallet';
 import { selectClaimForUri } from 'redux/selectors/claims';
 import WalletTipAmountSelector from './view';
+import { doOpenModal } from 'redux/actions/app';
 import { selectPreferredCurrency } from 'redux/selectors/settings';
 import { selectCanReceiveFiatTipsForUri, selectHasSavedCard } from 'redux/selectors/stripe';
 import { doTipAccountCheckForUri, doGetCustomerStatus } from 'redux/actions/stripe';
@@ -21,6 +22,7 @@ const select = (state, props) => {
 const perform = {
   doTipAccountCheckForUri,
   doGetCustomerStatus,
+  doOpenModal,
 };
 
 export default connect(select, perform)(WalletTipAmountSelector);
