+++ ui/component/walletTipAmountSelector/view.jsx
@@ -5,6 +5,7 @@ import { MINIMUM_PUBLISH_BID } from 'constants/claim';
 import { useIsMobile } from 'effects/use-screensize';
 import * as ICONS from 'constants/icons';
 import * as PAGES from 'constants/pages';
+import * as MODALS from 'constants/modal_types';
 import Button from 'component/button';
 import classnames from 'classnames';
 import React from 'react';
@@ -36,6 +37,8 @@ type Props = {
   preferredCurrency: string,
   doTipAccountCheckForUri: (uri: string) => void,
   doGetCustomerStatus: () => void,
+  doOpenModal: (modalId: string, modalProps?: any) => void,
+  modalProps?: any,
 };
 
 function WalletTipAmountSelector(props: Props) {
@@ -59,6 +62,8 @@ function WalletTipAmountSelector(props: Props) {
     preferredCurrency,
     doTipAccountCheckForUri,
     doGetCustomerStatus,
+    doOpenModal,
+    modalProps,
   } = props;
 
   const isMobile = useIsMobile();
@@ -284,7 +289,16 @@ function WalletTipAmountSelector(props: Props) {
         (!hasSavedCard
           ? getHelpMessage(
               <>
-                <Button navigate={`/$/${PAGES.SETTINGS_STRIPE_CARD}`} label={__('Add a Card')} button="link" />
+                <Button
+                  onClick={() =>
+                    doOpenModal(MODALS.ADD_CARD, {
+                      previousModal: modalProps ? MODALS.SEND_TIP : undefined,
+                      previousProps: modalProps,
+                    })
+                  }
+                  label={__('Add a Card')}
+                  button="link"
+                />
                 {' ' + __('To Tip Creators')}
               </>,
               'add-a-card-help-message'
