+++ ui/constants/modal_types.js
@@ -54,3 +54,4 @@ export const CONFIRM_ODYSEE_MEMBERSHIP = 'CONFIRM_ODYSEE_MEMBERSHIP';
 export const MEMBERSHIP_SPLASH = 'MEMBERSHIP_SPLASH';
 export const HIDE_RECOMMENDATION = 'HIDE_RECOMMENDATION';
 export const ANNOUNCEMENTS = 'ANNOUNCEMENTS';
+export const ADD_CARD = 'ADD_CARD';
