+++ ui/scss/component/_stripe-card.scss
@@ -433,7 +433,7 @@ pre {
   width: 284px;
   box-shadow: unset;
   color: #32325d;
-  background-color: white;
+  background-color: white !important;
 }
 
 .payment-details {
